Moved ModelInterface to Contracts
