+++ src/Models/Article.php
@@ -4,6 +4,7 @@
 namespace FlatFileCms\Models;
 
 use FlatFileCms\HtmlParser;
+use FlatFileCms\Models\Contracts\ModelInterface;
 use FlatFileCms\Models\Contracts\PublishInterface;
 use FlatFileCms\Models\Traits\Postable;
 use FlatFileCms\Models\Traits\Publishable;
