+++ src/Models/ContentBlock.php
@@ -2,6 +2,8 @@
 
 namespace FlatFileCms\Models;
 
+use FlatFileCms\Models\Contracts\ModelInterface;
+
 class ContentBlock extends Model implements ModelInterface
 {
     protected $folder = 'content_blocks';
