+++ src/Models/MetaTag.php
@@ -2,6 +2,8 @@
 
 namespace FlatFileCms\Models;
 
+use FlatFileCms\Models\Contracts\ModelInterface;
+
 class MetaTag extends Model implements ModelInterface
 {
     protected $folder = 'meta_tags';
